backend/requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.1
drf-spectacular>=0.27.2
dj-database-url>=2.2.0
django-cors-headers>=4.4.0
python-dotenv>=1.0.0
Pillow>=10.2.0
